added contig output to gff, refactored gff writing
